refactor package into several modules
