+++ recurring_ical_events/adapters/__init__.py
@@ -1,0 +1,1 @@
+"""All adapters for different kinds of components."""
