+++ recurring_ical_events/adapters/alarm.py
@@ -1,0 +1,13 @@
+from recurring_ical_events.adapters.component import ComponentAdapter
+
+
+from icalendar import Alarm
+
+
+class AbsoluteAlarmAdapter(ComponentAdapter):
+    """Adapter for absolute alarms."""
+
+    def __init__(self, alarm: Alarm, parent: ComponentAdapter):
+        """Create a new adapter."""
+        super().__init__(alarm)
+        self.parent = parent
