+++ recurring_ical_events/selection/__init__.py
@@ -1,0 +1,1 @@
+"""Select components for calculation."""
