+++ recurring_ical_events/series/__init__.py
@@ -1,0 +1,5 @@
+"""Calculation of occurrences in a series."""
+
+from .rrule import Series
+
+__all__ = ["Series"]
